Change the program structure. so inspector.h have more cpp files attach to it. but more clear which feature is in which file
