--- src/master/windows/inspector.cpp
@@ -1,978 +1,0 @@
-/*
- * Copyright (c) [2026] [Elly/Funique]
- *
- * This software is licensed under the [MIT License].
- * See the LICENSE file in the project root for more information.
-*/
-#include "inspector.h"
-#include <iostream>
-#include <string>
-#include <cstdlib>      // For std::getenv
-#include <filesystem>
-#include <stdexcept>
-
-namespace fs = std::filesystem;
-
-std::filesystem::path get_home_directory() {
-    const char* home_dir = nullptr;
-
-#if defined(_WIN32) || defined(_WIN64)
-    // Windows: Check USERPROFILE or HOMEDRIVE/HOMEPATH
-    home_dir = std::getenv("USERPROFILE");
-    if (!home_dir) {
-        const char* drive = std::getenv("HOMEDRIVE");
-        const char* path = std::getenv("HOMEPATH");
-        if (drive && path) {
-            return std::filesystem::path(drive) / path;
-        }
-    }
-#else
-    // Unix-like systems (Linux, macOS, etc.): Check HOME
-    home_dir = std::getenv("HOME");
-#endif
-
-    if (home_dir) {
-        return std::filesystem::path(home_dir);
-    } else {
-        // Fallback or throw an error if the home directory cannot be found
-        throw std::runtime_error("Could not find home directory");
-    }
-}
-
-std::vector<int32_t> Decode_Schedule_Time(int32_t value){
-    std::vector<int32_t> a = std::vector<int32_t>();
-    a.push_back(0); a.push_back(0); a.push_back(0);
-
-    if(value % 4 == 0) return std::vector<int32_t>();
-    if((value - 1) % 4 == 0 && value < 3073) a[0] = 1;
-    else if((value - 1) % 4 == 0 && value >= 3073) a[0] = 2;
-    else if((value + 1) % 4 == 0) a[0] = 3;
-
-    if(a[0] == 3){ // 24H
-        for(int32_t i = 0; i < 24; i++){ // 256 gap
-            int32_t starter = 3 + (256 * i);
-            if(value >= starter && value <= starter + 236){
-                a[1] = i;
-                a[2] = (value - starter) / 4;
-                break;
-            }
-        }
-    }else{ // 12H
-        for(int32_t i = 0; i < 12; i++){ // 256 gap
-            int32_t starter = -1;
-            if(a[0] == 2){ // PM
-                starter = 3073 + (256 * i);
-            }else{ // AM
-                starter = 1 + (256 * i);
-            }
-
-            if(value >= starter && value <= starter + 236){
-                a[1] = i;
-                a[2] = (value - starter) / 4;
-                break;
-            }
-        }
-    }
-
-    return a;
-}
-
-int32_t Encode_Schedule_Time(std::vector<int32_t> value){
-    if(value.size() == 0) return 0;
-    int32_t a = 0;
-    if(value.at(0) == 0) return 0;
-
-    if(value.at(0) == 1) a = 1; // AM
-    if(value.at(0) == 2) a = 3073; // PM
-    if(value.at(0) == 3) a = 3; // 24H
-
-    a += value.at(1) * 256;
-    a += value.at(2) * 4;
-    return a;
-}
-
-InspectorWindow::InspectorWindow(
-    std::shared_ptr<json> _setting, 
-    std::shared_ptr<GlobalState> _state, 
-    std::shared_ptr<GoProMaster> _master
-) 
-    : BaseWindow(_setting, _state, _master) {
-    title = "Inspector";
-    system_list_ordered = std::vector<int32_t>(GOPRO_SYSTEM_SETTING_SIZE);
-    video_setting_list_ordered = std::vector<int32_t>(GOPRO_VIDEO_SETTING_SIZE);
-    photo_setting_list_ordered = std::vector<int32_t>(GOPRO_PHOTO_SETTING_SIZE);
-    video_protune_list_ordered = std::vector<int32_t>(GOPRO_VIDEO_PROTUNE_SETTING_SIZE);
-    photo_protune_list_ordered = std::vector<int32_t>(GOPRO_PHOTO_PROTUNE_SETTING_SIZE);
-    status_list_ordered = std::vector<int32_t>(GOPRO_STATUS_SIZE);
-    reset_setting_order();
-    reset_status_order();
-}
-
-InspectorWindow::~InspectorWindow(){
-    
-}
-
-json InspectorWindow::get_window_data() {
-    json data = json::object();
-    data["put_finish"] = put_finish;
-    data["create_date_folder"] = create_date_folder;
-    data["current_download_location"] = state->current_download_location;
-    data["setting_order"] = json::object();
-    data["status_order"] = json::array();
-    data["setting_order"]["system_list_ordered"] = json::array();
-    data["setting_order"]["video_setting_list_ordered"] = json::array();
-    data["setting_order"]["photo_setting_list_ordered"] = json::array();
-    data["setting_order"]["video_protune_list_ordered"] = json::array();
-    data["setting_order"]["photo_protune_list_ordered"] = json::array();
-    for(int32_t i = 0; i < system_list_ordered.size(); i++){
-        data["setting_order"]["system_list_ordered"].push_back(system_list_ordered[i]);
-    }
-    for(int32_t i = 0; i < video_setting_list_ordered.size(); i++){
-        data["setting_order"]["video_setting_list_ordered"].push_back(video_setting_list_ordered[i]);
-    }
-    for(int32_t i = 0; i < photo_setting_list_ordered.size(); i++){
-        data["setting_order"]["photo_setting_list_ordered"].push_back(photo_setting_list_ordered[i]);
-    }
-    for(int32_t i = 0; i < video_protune_list_ordered.size(); i++){
-        data["setting_order"]["video_protune_list_ordered"].push_back(video_protune_list_ordered[i]);
-    }
-    for(int32_t i = 0; i < photo_protune_list_ordered.size(); i++){
-        data["setting_order"]["photo_protune_list_ordered"].push_back(photo_protune_list_ordered[i]);
-    }
-    for(int32_t i = 0; i < GOPRO_STATUS_SIZE; i++){
-        data["status_order"].push_back(status_list_ordered[i]);
-    }
-    return data;
-}
-
-void InspectorWindow::set_window_data(json data) {
-    if(data["put_finish"].is_boolean()){
-        put_finish = data["put_finish"].get<bool>();
-    }
-    if(data["create_date_folder"].is_boolean()){
-        create_date_folder = data["create_date_folder"].get<bool>();
-    }
-    if(data["current_download_location"].is_string()){
-        state->current_download_location = data["current_download_location"].get<std::string>();
-    }
-    if(data["setting_order"].is_object()){
-        if(data["setting_order"]["system_list_ordered"].is_array() && data["setting_order"]["system_list_ordered"].size() == system_list_ordered.size()){
-            for(int32_t i = 0; i < data["setting_order"]["system_list_ordered"].size(); i++){
-                if(data["setting_order"]["system_list_ordered"].at(i).is_number_integer()){
-                    system_list_ordered[i] = data["setting_order"]["system_list_ordered"].at(i).get<int32_t>();
-                }
-            }
-        }
-        if(data["setting_order"]["video_setting_list_ordered"].is_array() && data["setting_order"]["video_setting_list_ordered"].size() == video_setting_list_ordered.size()){
-            for(int32_t i = 0; i < data["setting_order"]["video_setting_list_ordered"].size(); i++){
-                if(data["setting_order"]["video_setting_list_ordered"].at(i).is_number_integer()){
-                    video_setting_list_ordered[i] = data["setting_order"]["video_setting_list_ordered"].at(i).get<int32_t>();
-                }
-            }
-        }
-        if(data["setting_order"]["photo_setting_list_ordered"].is_array() && data["setting_order"]["photo_setting_list_ordered"].size() == photo_setting_list_ordered.size()){
-            for(int32_t i = 0; i < data["setting_order"]["photo_setting_list_ordered"].size(); i++){
-                if(data["setting_order"]["photo_setting_list_ordered"].at(i).is_number_integer()){
-                    photo_setting_list_ordered[i] = data["setting_order"]["photo_setting_list_ordered"].at(i).get<int32_t>();
-                }
-            }
-        }
-        if(data["setting_order"]["video_protune_list_ordered"].is_array() && data["setting_order"]["video_protune_list_ordered"].size() == video_protune_list_ordered.size()){
-            for(int32_t i = 0; i < data["setting_order"]["video_protune_list_ordered"].size(); i++){
-                if(data["setting_order"]["video_protune_list_ordered"].at(i).is_number_integer()){
-                    video_protune_list_ordered[i] = data["setting_order"]["video_protune_list_ordered"].at(i).get<int32_t>();
-                }
-            }
-        }
-        if(data["setting_order"]["photo_protune_list_ordered"].is_array() && data["setting_order"]["photo_protune_list_ordered"].size() == photo_protune_list_ordered.size()){
-            for(int32_t i = 0; i < photo_protune_list_ordered.size(); i++){
-                if(data["setting_order"]["photo_protune_list_ordered"].at(i).is_number_integer()){
-                    photo_protune_list_ordered[i] = data["setting_order"]["photo_protune_list_ordered"].at(i).get<int32_t>();
-                }
-            }
-        }
-    }
-    if(data["status_order"].is_array() && data["status_order"].size() == GOPRO_STATUS_SIZE){
-        for(int32_t i = 0; i < GOPRO_STATUS_SIZE; i++){
-            if(data["status_order"].at(i).is_number_integer()){
-                status_list_ordered[i] = data["status_order"].at(i).get<int32_t>();
-            }
-        }
-    }
-}
-
-void InspectorWindow::render(){
-    ImGui::Begin("Inspector", &enable, w_flag);
-    {
-        std::lock_guard<std::mutex> lock(master->camera_mtx);
-        int32_t camera_ip = master->findCamera(state->current_camera_item);
-        bool should_disabled = state->current_camera_item.size() < 10 || camera_ip == -1 || !state->current_setting_items_bind;
-
-        draw_header();
-        ImGui::BeginDisabled(should_disabled);
-
-        ImGui::Separator();
-
-        if(ImGui::BeginTabBar("Inspector_Bar##Top")){
-            if(ImGui::BeginTabItem("Setting##Inspector_Bar_Item")){
-                if(ImGui::Button("Reset Setting Order")){
-                    reset_setting_order();
-                    state->update_server();
-                }
-                if(ImGui::BeginTabBar("Inspector_Bar##Second")){
-                    if(ImGui::BeginTabItem("System##Inspector_Bar_Item")){
-                        draw_system();
-                        ImGui::EndTabItem();
-                    }
-                    if(ImGui::BeginTabItem("Setting##Inspector_Bar_Item")){
-                        draw_setting();
-                        ImGui::EndTabItem();
-                    }
-                    if(ImGui::BeginTabItem("Protune##Inspector_Bar_Item")){
-                        draw_protune();
-                        ImGui::EndTabItem();
-                    }
-                    ImGui::EndTabBar();
-                }
-                ImGui::EndTabItem();
-            }
-            if(ImGui::BeginTabItem("Status##Inspector_Bar_Item")){
-                if(ImGui::Button("Reset Status Order")){
-                    reset_status_order();
-                    state->update_server();
-                }
-                draw_status();
-                ImGui::EndTabItem();
-            }
-            if(ImGui::BeginTabItem("Media##Inspector_Bar_Item")){
-                draw_media();
-                ImGui::EndTabItem();
-            }
-            ImGui::EndTabBar();
-        }
-        
-        ImGui::EndDisabled();
-    }
-    ImGui::End();
-}
-
-void InspectorWindow::draw_header(){
-    ImGui::InputText("Camera Name", &state->current_camera_name);
-    if(ImGui::Button("Rename Camera")){
-        master->command_with_value("rename", state->current_camera_item, state->current_camera_name);
-    }
-    ImGui::Text("Name: %s, IP: %s", state->current_camera_name.c_str(), state->current_camera_item.c_str());
-}
-
-void InspectorWindow::draw_system(){
-    _draw_setting(system_list_ordered);
-}
-
-void InspectorWindow::draw_setting(){
-    if (state->current_status_items[std::to_string(PRESET_ID)].is_number()) {
-        int32_t preset = state->current_status_items[std::to_string(PRESET_ID)].get<int32_t>();
-        if(preset == 0){
-            _draw_setting(video_setting_list_ordered);
-        }else{
-            _draw_setting(photo_setting_list_ordered);
-        }
-    }
-}
-
-void InspectorWindow::draw_protune(){
-    if (state->current_status_items[std::to_string(PRESET_ID)].is_number()) {
-        int32_t preset = state->current_status_items[std::to_string(PRESET_ID)].get<int32_t>();
-        if(preset == 0){
-            _draw_setting(video_protune_list_ordered);
-        }else{
-            _draw_setting(photo_protune_list_ordered);       
-        }
-    }
-}
-
-void InspectorWindow::draw_status(){
-    int move_from = -1, move_to = -1;
-    for(int32_t i = 0; i < GOPRO_STATUS_SIZE; i++){
-        float x = ImGui::GetCursorPosX();
-        int32_t id = status_list_ordered[i];
-        const int32_t type = GET_STATUS_TYPE_BY_ID(id);
-        std::string name = GET_STATUS_NAME_BY_ID(id);
-        if (name.size() == 0) continue;
-
-        name += "##InspectorTitle_status";
-        std::string pp = "##InspectorTitle_status_option_" + std::string(GET_STATUS_NAME_BY_ID(id));
-
-        if(type == (int32_t)CAMERA_STATUS_TYPE::OPTION){
-            size_t size = GET_STATUS_SIZE_BY_ID(id);
-            if (!state->current_status_items[std::to_string(id)].is_number()) continue;
-            int32_t select_index = state->current_status_items[std::to_string(id)].get<int32_t>();
-            const char** select_string_list = GET_STATUS_STRING_BY_ID(id);
-            std::string select_value_display = select_string_list[select_index];
-            pp = std::to_string(select_index) + " " + select_value_display + pp;
-        }else if (type == (int32_t)CAMERA_STATUS_TYPE::BOOLEAN){
-            if (!state->current_status_items[std::to_string(id)].is_number()) continue;
-            int32_t value = state->current_status_items[std::to_string(id)].get<int32_t>();
-            std::string display = "";
-            if(value == 1) display = "true";
-            else display = "false";
-            pp = display + pp;
-        }else if (type == (int32_t)CAMERA_STATUS_TYPE::INT){
-            if (!state->current_status_items[std::to_string(id)].is_number()) continue;
-            int32_t value = state->current_status_items[std::to_string(id)].get<int32_t>();
-            pp = std::to_string(value) + pp;
-        }else if (type == (int32_t)CAMERA_STATUS_TYPE::LONG){
-            if (!state->current_status_items[std::to_string(id)].is_number()) continue;
-            int64_t value = state->current_status_items[std::to_string(id)].get<int64_t>();
-            pp = std::to_string(value) + pp;
-        }else if (type == (int32_t)CAMERA_STATUS_TYPE::STRING){
-            if (!state->current_status_items[std::to_string(id)].is_string()) continue;
-            std::string value = state->current_status_items[std::to_string(id)].get<std::string>();
-            pp = value + pp;
-        }else {
-            std::cout << "Type is unknown: " << type << std::endl;
-            continue;   
-        }
-
-        ImGui::Selectable(pp.c_str(), false, ImGuiSelectableFlags_SpanAllColumns);
-
-        ImGuiDragDropFlags src_flags = 0;
-        src_flags |= ImGuiDragDropFlags_SourceNoDisableHover;
-        src_flags |= ImGuiDragDropFlags_SourceNoHoldToOpenOthers;
-        if(ImGui::BeginDragDropSource(src_flags)){
-            if (!(src_flags & ImGuiDragDropFlags_SourceNoPreviewTooltip))
-                ImGui::Text("Moving \"%s\"", name.c_str());
-            ImGui::SetDragDropPayload("INSPECTOR_STATUS", &i, sizeof(int));
-            ImGui::EndDragDropSource();
-        }
-
-        if(ImGui::BeginDragDropTarget()){
-            ImGuiDragDropFlags target_flags = 0;
-            target_flags |= ImGuiDragDropFlags_AcceptBeforeDelivery;
-            target_flags |= ImGuiDragDropFlags_AcceptNoDrawDefaultRect;
-            if (const ImGuiPayload* payload = ImGui::AcceptDragDropPayload("INSPECTOR_STATUS", target_flags))
-            {
-                move_from = *(const int*)payload->Data;
-                move_to = i;
-            }
-            ImGui::EndDragDropTarget();
-        }
-        ImGui::SameLine();
-        ImGui::SetCursorPosX(x + ImGui::CalcItemWidth() + ImGui::GetStyle().ItemInnerSpacing.x);
-        ImGui::TextUnformatted(GET_STATUS_NAME_BY_ID(id));
-    }
-    if (move_from != -1 && move_to != -1)
-    {
-        const int32_t tmp = status_list_ordered[move_from];
-        status_list_ordered[move_from] = status_list_ordered[move_to];
-        status_list_ordered[move_to] = tmp;
-        state->update_server();
-    }
-}
-
-void InspectorWindow::draw_media(){
-    ImGuiStyle& style = ImGui::GetStyle();
-    ImVec2 size = ImGui::GetWindowSize();
-    ImVec2 button_size = ImVec2(size.x / 2.0F - style.ItemSpacing.x, 0);
-
-    int32_t camera_ip = master->findCamera(state->current_camera_item);
-    if(ImGui::InputText("Media Download", &state->current_download_location)){
-        state->update_server();
-    }
-    if(ImGui::Button("All Download", button_size)){
-        std::string buffer = state->current_download_location;
-        while(buffer.size() > 0 && buffer.at(buffer.size() - 1) == '/'){
-            buffer.pop_back();
-        }
-        if(master->directoryExists(buffer)){
-            if(create_date_folder){
-                std::string date = getCurrentDateTimeString();
-                buffer.append("/" + date);
-            }
-            fs::create_directories(buffer);
-            master->download_last_media(buffer, put_finish);
-        }
-    }
-    if(ImGui::IsItemHovered()) ImGui::SetTooltip("Download all exist camera instances");
-    ImGui::SameLine();
-    if(ImGui::Button("Single Download", button_size)){
-        
-    }
-    if(ImGui::IsItemHovered()) ImGui::SetTooltip("Download current select camera instance");
-
-    if(ImGui::Button("Open Home", button_size)){
-        std::system((std::string("open \"") + get_home_directory().string() + std::string("\"")).c_str());
-    }
-    if(ImGui::IsItemHovered()) ImGui::SetTooltip("Open file explorer for home directory");
-    ImGui::SameLine();
-    if(ImGui::Button("Open Select Path", button_size)){
-        if(fs::exists(state->current_download_location)){
-            std::system((std::string("open \"") + state->current_download_location + std::string("\"")).c_str());
-        }
-    }
-
-    if(ImGui::Checkbox("Create Date Folder", &create_date_folder)){
-        state->update_server();
-    }
-    if(ImGui::Checkbox("Added Finish File", &put_finish)){
-        state->update_server();
-    }
-
-    if(ImGui::IsItemHovered()) ImGui::SetTooltip("Open file explorer for path select directory");
-    if(camera_ip >= 0){
-        std::shared_ptr<CameraInfo> t = master->getCameras()[camera_ip];
-        ImGui::LabelText("Server", "%s", t->server.c_str());
-        ImGui::LabelText("Last Media", "%s", t->last_media.c_str());
-    }
-}
-
-void InspectorWindow::draw_command_local(){
-    std::lock_guard<std::mutex> lock(master->camera_mtx);
-    ImGui::Text("Single Camera Control");
-
-    int32_t current_camera = master->findCamera(state->current_camera_item);
-    bool should_disabled = state->current_camera_item.size() < 10 || current_camera == -1;
-    ImGui::BeginDisabled(should_disabled);
-
-    ImGuiStyle& style = ImGui::GetStyle();
-    ImVec2 size = ImGui::GetWindowSize();
-    ImVec2 button_2size = ImVec2(size.x / 2.0F - style.ItemSpacing.x, 0);
-    ImVec2 button_4size = ImVec2(size.x / 4.0F - style.ItemSpacing.x, 0);
-    ImVec2 full_button_size = ImVec2(size.x - style.ItemSpacing.x, 0);
-
-    if(current_camera != -1) {
-        CameraInfo info = *(master->getCameras().at(current_camera));
-
-        if(ImGui::Button("Record", button_2size)) master->command_only(info.server, "shutter_on", info.ip); ImGui::SameLine();
-        if(ImGui::Button("Stop", button_2size)) master->command_only(info.server, "shutter_off", info.ip);
-
-        if(ImGui::Button("Connect", button_4size)) master->command_only(info.server, "usb_on", info.ip); ImGui::SameLine();
-        if(ImGui::Button("Disconnect", button_4size)) master->command_only(info.server, "usb_off", info.ip); ImGui::SameLine();
-        if(ImGui::Button("Shutdown", button_4size)) master->command_only(info.server, "usb_on", info.ip); ImGui::SameLine();
-        int32_t lo = master->haslocate(info.server, info.ip);
-        if(lo == -1){
-            if(ImGui::Button("Locate", button_4size)) master->locate(info.server, info.ip);
-        }else{
-            if(ImGui::Button("UnLocate", button_4size)) master->locate(info.server, info.ip);
-        }
-
-        ImGui::Dummy(ImVec2(0, 20));
-        ImGui::Separator();
-        ImGui::Dummy(ImVec2(0, 20));
-
-        if(state->current_status_items[std::to_string(PRESET_ID)].is_number_integer()){
-
-            int32_t cuccrent_select = state->current_status_items[std::to_string(PRESET_ID)].get<int32_t>();
-            int32_t index = -1;
-            for(int32_t i = 0; i < GOPRO_MODE_SIZE; i++){
-                if(GOPRO_MODE_VALUE[i] == cuccrent_select){
-                    index = i;
-                    break;
-                }
-            }
-
-            std::string display_mode = GOPRO_MODE_STRING[index];
-            
-            if(ImGui::BeginCombo("Mode##SCC", display_mode.c_str())){
-                for (int n = 0; n < GOPRO_MODE_SIZE; n++)
-                {
-                    std::string curr = GOPRO_MODE_STRING[n];
-                    std::string curr_b = curr + "##SCC_Option";
-                    bool is_selected = (state->current_mode_item_string == curr); // You can store your selection however you want, outside or inside your objects
-                    if (ImGui::Selectable(curr_b.c_str(), is_selected)){
-                        master->presetSwitch("", info.ip, GOPRO_MODE_VALUE[n]);
-                        state->current_mode_item_string = curr;
-                        state->current_mode_item = n;
-                    }
-                    if (is_selected)
-                        ImGui::SetItemDefaultFocus();   // You may set the initial focus when opening the combo (scrolling + for keyboard navigation support)
-                }
-                ImGui::EndCombo();
-            }
-        }
-        
-        if(ImGui::Button("Setting Apply All By ID", full_button_size)) {
-            int32_t v = state->current_setting_items[std::to_string(state->apply_all_item)].get<int32_t>();
-            master->apply("", state->apply_all_item, GET_SETTING_VALUE_BY_ID(state->apply_all_item)[v]);
-        }
-
-        if(ImGui::BeginCombo("ID", state->apply_all_item_string.c_str())){
-            for (int n = 0; n < GOPRO_SETTING_SIZE; n++)
-            {
-                int32_t id = GOPRO_SETTING_IDS[n];
-                std::string curr = GET_SETTING_NAME_BY_ID(id);
-                bool is_selected = (state->current_mode_item_string == curr); // You can store your selection however you want, outside or inside your objects
-                std::string curr_b = curr + "##Setting_ID_Option";
-                if (ImGui::Selectable(curr_b.c_str(), is_selected)){
-                    state->apply_all_item_string = curr;
-                    state->apply_all_item = GOPRO_SETTING_IDS[n];
-                }
-                if (is_selected)
-                    ImGui::SetItemDefaultFocus();   // You may set the initial focus when opening the combo (scrolling + for keyboard navigation support)
-            }
-            ImGui::EndCombo();
-        }
-
-        json buffer_c = json::object();
-        std::string try_apply = "";
-        if(master->getSettingsFromCamera(info, buffer_c)){
-            if(buffer_c[std::to_string(state->apply_all_item)].is_number_integer()){
-                int32_t v = buffer_c[std::to_string(state->apply_all_item)].get<int32_t>();
-                try_apply = GET_SETTING_STRING_BY_ID(state->apply_all_item)[v];
-                ImGui::Text("Trying Apply Value: %s", try_apply.c_str());
-            }
-        }
-
-        if (ImGui::BeginTable("Local_Apply", 4, ImGuiTableFlags_Borders | ImGuiTableFlags_RowBg)) {
-            ImGui::TableSetupColumn("Server Address##Local_Apply_header");
-            ImGui::TableSetupColumn("IP Address##Local_Apply_header");
-            ImGui::TableSetupColumn("Value##Local_Apply_header");
-            ImGui::TableSetupColumn("OK##Local_Apply_header");
-            ImGui::TableHeadersRow();
-
-            for(auto& camera: master->getCameras()){
-                json buffer = json::object();
-                if(master->getSettingsFromCamera(*camera, buffer)){
-                    if(buffer[std::to_string(state->apply_all_item)].is_number_integer()){
-                        int32_t v = buffer[std::to_string(state->apply_all_item)].get<int32_t>();
-                        std::string c_v = GET_SETTING_STRING_BY_ID(state->apply_all_item)[v];
-                        bool eq = c_v == try_apply;
-                        std::string ok = "";
-                        if(eq) ok = "O";
-                        else ok = "X";
-                        ImGui::TableNextRow();
-                        ImGui::TableSetColumnIndex(0);
-                        ImGui::Text("%s", camera->server.c_str());
-                        ImGui::TableSetColumnIndex(1);
-                        ImGui::Text("%s", camera->ip.c_str());
-                        ImGui::TableSetColumnIndex(2);
-                        ImGui::Text("%s", c_v.c_str());   
-                        ImGui::TableSetColumnIndex(3);
-                        ImGui::Text("%s", ok.c_str());
-                    }
-                }
-            }
-            ImGui::EndTable();
-        }
-    }
-    ImGui::EndDisabled();
-}
-
-void InspectorWindow::draw_command_global(){
-    ImGui::Text("Global Controls");
-    ImGui::Text("Commands applied to all connected cameras");
-
-    ImGuiStyle& style = ImGui::GetStyle();
-    ImVec2 size = ImGui::GetWindowSize();
-    ImVec2 button_size = ImVec2(size.x / 2.0F - style.ItemSpacing.x, 0);
-    ImVec2 full_button_size = ImVec2(size.x - style.ItemSpacing.x, 0);
-    
-    if(ImGui::Button("Scan All", button_size)) master->command_only("scan"); ImGui::SameLine();
-    if(ImGui::IsItemHovered()) ImGui::SetTooltip("Scan all websocket server for cameras");
-    if(ImGui::Button("Scan Server", button_size)) state->command_sender("scan_server");
-    if(ImGui::IsItemHovered()) ImGui::SetTooltip("Scan one websocket server for cameras\nThis will popup a window for you to enter websocket server address to targeting");
-
-    if(ImGui::Button("Add Camera", button_size)) state->command_sender("add_camera"); ImGui::SameLine();
-    if(ImGui::IsItemHovered()) ImGui::SetTooltip("Menually add camera address to one websocket server list");
-    if(ImGui::Button("Clean Camera", button_size)) master->command_only("clean");
-    if(ImGui::IsItemHovered()) ImGui::SetTooltip("Clean all websocket server camera record");
-
-    if(ImGui::Button("Connect All", button_size)) master->command_only("usb_on"); ImGui::SameLine();
-    if(ImGui::IsItemHovered()) ImGui::SetTooltip("Tells all cameras usb control on");
-    if(ImGui::Button("Disconnect All", button_size)) master->command_only("usb_off");
-    if(ImGui::IsItemHovered()) ImGui::SetTooltip("Tells all cameras usb control off");
-
-    if(ImGui::Button("Reboot All", button_size)) master->command_only("reboot"); ImGui::SameLine();
-    if(ImGui::IsItemHovered()) ImGui::SetTooltip("Tells all cameras reboot right now");
-    if(ImGui::Button("Shutdown All", button_size)) master->command_only("shutdown");
-    if(ImGui::IsItemHovered()) ImGui::SetTooltip("Tells all cameras shutdown right now");
-
-    if(ImGui::Button("Record All", button_size)) master->command_only("shutter_on"); ImGui::SameLine();
-    if(ImGui::Button("Stop All", button_size)) master->command_only("shutter_off");
-
-    if(ImGui::Button("Enter Webcam", button_size)) master->webcam_only("preview"); ImGui::SameLine();
-    if(ImGui::Button("Exit Webcam", button_size)) master->webcam_only("exit");
-
-    if(ImGui::Button("Start Webcam", button_size)) state->command_sender("start_webcam"); ImGui::SameLine();
-    if(ImGui::Button("Sync Time", button_size)) master->command_only("datetime");
-
-    if(ImGui::Button("Last Media", button_size)) master->media_only("lastmedia"); ImGui::SameLine();
-    if(ImGui::Button("Close Preview", button_size)) master->preview_end("", "");
-}
-
-void InspectorWindow::reset_setting_order(){
-    for(int32_t i = 0; i < system_list_ordered.size(); i++){
-        system_list_ordered[i] = GOPRO_SYSTEM_SETTING_IDS[i];
-    }
-    for(int32_t i = 0; i < video_setting_list_ordered.size(); i++){
-        video_setting_list_ordered[i] = GOPRO_VIDEO_SETTING_IDS[i];
-    }
-    for(int32_t i = 0; i < photo_setting_list_ordered.size(); i++){
-        photo_setting_list_ordered[i] = GOPRO_PHOTO_SETTING_IDS[i];
-    }
-    for(int32_t i = 0; i < video_protune_list_ordered.size(); i++){
-        video_protune_list_ordered[i] = GOPRO_VIDEO_PROTUNE_SETTING_IDS[i];
-    }
-    for(int32_t i = 0; i < photo_protune_list_ordered.size(); i++){
-        photo_protune_list_ordered[i] = GOPRO_PHOTO_PROTUNE_SETTING_IDS[i];
-    }
-}
-
-void InspectorWindow::reset_status_order(){
-    for(int32_t i = 0; i < GOPRO_STATUS_SIZE; i++){
-        status_list_ordered[i] = GOPRO_STATUS_IDS[i];
-    }
-}
-
-void InspectorWindow::_draw_setting(std::vector<int32_t>& ordered){
-    int32_t move_from = -1, move_to = -1;
-    int32_t model_enum = _get_current_model();
-    for(int32_t i = 0; i < ordered.size(); i++){
-        int32_t id = ordered[i];
-        std::string name = GET_SETTING_NAME_BY_ID(id);
-        size_t size = GET_SETTING_SIZE_BY_ID(id);
-        int32_t ava = GET_SETTING_AVA_BY_ID(id);
-        if((model_enum&ava) == 0){
-            continue;
-        }
-        if (!state->current_setting_items[std::to_string(id)].is_number()) {
-            continue;
-        }
-        if (name.size() == 0) {
-            continue;
-        }
-        if(!conditional_filter(model_enum, id)){
-            continue;
-        }
-
-        name += "##InspectorTitle";
-        int32_t select_index = state->current_setting_items[std::to_string(id)].get<int32_t>();
-        const char** select_string_list = GET_SETTING_STRING_BY_ID(id);
-        if(select_string_list == nullptr) {
-            std::cerr << "Inspector: select_string_list == nullptr" << std::endl;
-            continue;
-        }
-        if(select_index >= size) {
-            std::cerr << "Inspector: select_index >= size..." << id << "..." << select_index << "..." << size << std::endl;
-            continue;
-        }
-        const char* select_string = select_string_list[select_index];
-        if(select_string == nullptr) continue;
-
-        const int32_t* values_id = GET_SETTING_VALUE_BY_ID(id);
-        const int32_t* support_id = GET_SETTING_SUPPORT_BY_ID(id);
-
-        if(ImGui::BeginCombo(name.c_str(), select_string)){
-            for (int n = 0; n < size; n++)
-            {
-                int32_t supp = support_id[n];
-                if((model_enum&supp) == 0){
-                    continue;
-                }
-                if(!conditional_filter_option(model_enum, id, n)){
-                    continue;
-                }
-
-                std::string option = GET_SETTING_STRING_BY_ID(id)[n];
-                if(option.size() == 0) continue;
-                bool is_selected = (state->current_setting_items[std::to_string(id)] == n); // You can store your selection however you want, outside or inside your objects
-                option += ("##InspectorOption_" + name); 
-                if (ImGui::Selectable(option.c_str(), is_selected))
-                {
-                    state->current_setting_items[std::to_string(id)] = n; // Change index
-                    master->apply(state->current_camera_item, id, values_id[n]);
-                }
-                if (is_selected)
-                    ImGui::SetItemDefaultFocus();   // You may set the initial focus when opening the combo (scrolling + for keyboard navigation support)
-            }
-            ImGui::EndCombo();
-        }
-        
-        ImGuiDragDropFlags src_flags = 0;
-        src_flags |= ImGuiDragDropFlags_SourceNoDisableHover;
-        src_flags |= ImGuiDragDropFlags_SourceNoHoldToOpenOthers;
-        //src_flags |= ImGuiDragDropFlags_SourceNoPreviewTooltip;
-        if(ImGui::BeginDragDropSource(src_flags)){
-            if (!(src_flags & ImGuiDragDropFlags_SourceNoPreviewTooltip))
-                ImGui::Text("Moving \"%s\"", name.c_str());
-            ImGui::SetDragDropPayload("INSPECTOR_SETTING", &i, sizeof(int));
-            ImGui::EndDragDropSource();
-        }
-
-        if(ImGui::BeginDragDropTarget()){
-            ImGuiDragDropFlags target_flags = 0;
-            target_flags |= ImGuiDragDropFlags_AcceptBeforeDelivery;
-            target_flags |= ImGuiDragDropFlags_AcceptNoDrawDefaultRect;
-            if (const ImGuiPayload* payload = ImGui::AcceptDragDropPayload("INSPECTOR_SETTING", target_flags))
-            {
-                move_from = *(const int*)payload->Data;
-                move_to = i;
-            }
-            ImGui::EndDragDropTarget();
-        }
-    }
-    if (move_from != -1 && move_to != -1)
-    {
-        const int32_t tmp = ordered[move_from];
-        ordered[move_from] = ordered[move_to];
-        ordered[move_to] = tmp;
-        //ImGui::SetDragDropPayload("INSPECTOR_SETTING", &move_to, sizeof(int));
-        state->update_server();
-    }
-}
-
-int32_t InspectorWindow::_get_current_model(){
-    if(state->current_hw_items["model_name"].is_string()){
-        std::string model_name = state->current_hw_items["model_name"].get<std::string>();
-        if(model_name == "MAX 2") return MODEL_MAX2;
-        else if(model_name == "HERO13 Black") return MODEL_13;
-        else if(model_name == "HERO12 Black") return MODEL_12;
-        else if(model_name == "HERO11 Black") return MODEL_11_BLACK;
-        else if(model_name == "HERO11 Black Mini") return MODEL_11;
-        else if(model_name == "HERO10 Black") return MODEL_10;
-        else if(model_name == "HERO9 Black") return MODEL_9;
-    }
-    else return 0;
-}
-
-bool InspectorWindow::conditional_filter(int32_t mymodel, int32_t setting_id){
-    int32_t preset = state->try_get_status_int32_by_id(PRESET_ID);
-
-    if(setting_id == SHUTTER_SPEED_VIDEO_ID){
-        int32_t profile = state->try_get_setting_int32_by_id(PROFILES_ID);
-        if(profile == 1) return false; // HDR
-    }
-    else if(setting_id == EXPOSURE_ID){
-        int32_t shutter = state->try_get_setting_int32_by_id(SHUTTER_SPEED_VIDEO_ID);
-        if(shutter != 0 && preset == 0) return false; // Auto, and video mode
-    }
-    return true;
-}
-
-bool InspectorWindow::conditional_filter_option(int32_t mymodel, int32_t setting_id, int32_t value_index){
-    int32_t preset = state->try_get_status_int32_by_id(PRESET_ID);
-
-    int32_t profile = state->try_get_setting_int32_by_id(PROFILES_ID);
-    int32_t res = state->try_get_setting_int32_by_id(VIDEO_RESOLUTION_ID);
-    int32_t aspect = state->try_get_setting_int32_by_id(VIDEO_ASPECT_RATIO_ID);
-    int32_t fps = state->try_get_setting_int32_by_id(FRAMES_PER_SECOND_ID);
-    int32_t antif = state->try_get_setting_int32_by_id(ANTI_FLICKER_V2_ID);
-    int32_t videolen = state->try_get_setting_int32_by_id(VIDEO_LENS_ID);
-    
-    int32_t res_id = VIDEO_RESOLUTION_VALUE[res];
-    int32_t aspect_id = VIDEO_ASPECT_RATIO_VALUE[aspect];
-    int32_t fps_id = FRAMES_PER_SECOND_VALUE[fps];
-    int32_t videolen_id = VIDEO_LENS_VALUE[videolen];
-
-    if(setting_id == VIDEO_RESOLUTION_ID){
-        int32_t value_id = VIDEO_RESOLUTION_VALUE[value_index];
-        if(preset == 0){
-            if(profile == 0){ // Standard
-                if(aspect_id == 0){ // "4:3"
-                    if(value_id != 6) return false;
-                }
-                else if(aspect_id == 1){ // "16:9"
-                    if(fps_id == 0){ // 240
-                        // 2.7K, 1080
-                        if(value_id != 4 && value_id != 9) return false;
-                    }
-                    if(fps_id == 1){ // 120
-                        // 1080, 4K
-                        if(value_id != 9 && value_id != 1) return false;
-                    }
-                    if(fps_id == 5){ // 60
-                        if(videolen_id == 9){ // Hyperview
-                            // 4K
-                            if(value_id != 1) return false;
-                        }else{
-                            // 1080, 4K, 5.3K
-                            if(value_id != 9 && value_id != 1 && value_id != 100) return false;
-                        }
-                    }
-                    if(fps_id == 8){ // 30
-                        // 1080, 4K, 5.3K
-                        if(value_id != 9 && value_id != 1 && value_id != 100) return false;
-                    }
-                    else if(fps_id == 10){ // 24
-                        // 4K, 5.3K
-                        if(value_id != 1 && value_id != 100) return false;
-                    }
-                }
-                else if(aspect_id == 3){ // "8:7"
-                    // 60
-                    if(fps_id == 5){
-                        // 4K 8:7
-                        if(value_id != 108) return false;
-                    }else{
-                        // 4K 8:7, 5.3K 8:7
-                        if(value_id != 107 && value_id != 108) return false;
-                    }
-                }
-                else if(aspect_id == 4){ // "9:16"
-                    // 1080 9:16, 4K 9:16
-                    if(value_id != 109 && value_id != 110) return false;
-                }
-            }
-            else if(profile == 1){ // HDR
-                if(aspect_id == 3){ // "8:7"
-                    // 4K 8:7
-                    if(value_id != 108) return false;
-                }
-                else if(aspect_id == 1){ // "16:9"
-                    if(fps_id == 5 || videolen_id == 9) { // 60
-                        // 4K
-                        if(value_id != 1) return false;
-                    }else{
-                        // 4K, 5.3K
-                        if(value_id != 1 && value_id != 100) return false;
-                    }
-                }
-            }
-            else if(profile == 2){ // LOG
-                if(aspect_id == 3){ // "8:7"
-                    if(fps_id == 5){ // 60
-                        // 4K 8:7
-                        if(value_id != 108) return false;
-                    }else{
-                        // 5.3K 8:7, 4K 8:7
-                        if(value_id != 107 && value_id != 108) return false;
-                    }
-                }
-                else if(aspect_id == 1){ // "16:9"
-                    if(fps_id == 1 || videolen_id == 9){ // 120
-                        // 4K
-                        if(value_id != 1) return false;
-                    }else{
-                        // 4K, 5.3K
-                        if(value_id != 1 && value_id != 100) return false;
-                    }
-                }
-            }
-        }
-    }
-    else if(setting_id == VIDEO_ASPECT_RATIO_ID){
-        int32_t value_id = VIDEO_ASPECT_RATIO_VALUE[value_index];
-        if(preset == 0){ // Video
-            if(profile == 0){ // Standard
-                if(value_id > 4) return false;
-            }
-            else if(profile == 1){ // HDR
-                if(value_id != 1 && value_id != 3) return false;
-            }
-            else if(profile == 2){ // LOG
-                if(value_id != 1 && value_id != 3) return false;
-            }
-        }
-    }
-    else if(setting_id == FRAMES_PER_SECOND_ID){
-        int32_t value_id = FRAMES_PER_SECOND_VALUE[value_index];
-        if(preset == 0){ // Video
-            if(profile == 0) { // Standard
-                if(aspect_id == 0){ // "4:3"
-                    // 120
-                    if(value_id != 1) return false;
-                }
-                else if(aspect_id == 1){ // "16:9"
-                    if(res_id == 1){ // 4K
-                        // 30, 60, 120, 24
-                        if(value_id != 8 && value_id != 5 && value_id != 1 && value_id != 10) return false;
-                    }
-                    else if(res_id == 100){ // 5.3K
-                    }
-                    else if(res_id == 4){ // 2.7K
-                        // 240
-                        if(value_id != 0) return false;
-                    }else{ // 1080
-                        // 30, 60, 120, 240
-                        if(value_id != 8 && value_id != 5 && value_id != 1 && value_id != 0) return false;
-                    }
-                }
-                else if(aspect_id == 3){ // "8:7"
-                    if(res_id == 107){ // 5.3K 8:3
-                        // 24, 30
-                        if(value_id != 10 && value_id != 8) return false;
-                    }else{ // 4K 8:7
-                        // 24, 30, 60
-                        if(value_id != 10 && value_id != 8 && value_id != 5) return false;
-                    }
-                }
-                else if(aspect_id == 4){ // "9:16"
-                    // 60, 30
-                    if(value_id != 5 && value_id != 8) return false;
-                }
-            }
-            else if(profile == 1) { // HDR
-                if(aspect_id == 3){ // 8:7
-                    // 24, 30
-                    if(value_id != 10 && value_id != 8) return false;
-                }else{ // 16:9
-                    if(res_id == 100 || videolen_id == 9 || videolen_id == 3){ // 5.3K
-                        // 24, 30
-                        if(value_id != 10 && value_id != 8) return false;
-                    }else{ // 4K
-                        // 24, 30, 60
-                        if(value_id != 10 && value_id != 8 && value_id != 5) return false;
-                    }
-                }
-            }
-            else if(profile == 2) { // LOG
-                if(aspect_id == 1){ // 16:9
-                    if(res_id == 100){ // 5.3K
-                        if(videolen_id == 9) { // Hyperview
-                            // 30, 24
-                            if(value_id != 8 && value_id != 10) return false;
-                        }else{
-                            // 30, 60, 24
-                            if(value_id != 8 && value_id != 5 && value_id != 10) return false;
-                        }
-                    }else{ // 4K
-                        if(videolen_id == 9) { // Hyperview
-                            // 30, 60, 24
-                            if(value_id != 8 && value_id != 5 && value_id != 10) return false;
-                        }else{
-                            // 30, 60, 120, 24
-                            if(value_id != 8 && value_id != 5 && value_id != 1 && value_id != 10) return false;
-                        }
-                    }
-                }
-                else if(aspect_id == 3){ // 8:7
-                    // 30, 60, 24
-                    if(value_id != 8 && value_id != 5 && value_id != 10) return false;                    
-                }
-            }
-        }
-    }
-    else if(setting_id == VIDEO_LENS_ID){
-        int32_t value_id = VIDEO_LENS_VALUE[value_index];
-        if(preset == 0){ // video
-
-        }
-    }
-    else if(setting_id == SHUTTER_SPEED_VIDEO_ID){
-        if((mymodel&(ANTI_FLICKER_V2_AVA)) > 0){
-            int32_t value_id = SHUTTER_SPEED_VIDEO_VALUE[value_index];
-            if(antif == 1){
-                if(value_id != 0 && value_id != 51 && value_id != 50 &&
-                value_id != 30 && value_id != 49 && value_id != 29 &&
-                value_id != 48 && value_id != 28 && value_id != 47 &&
-                value_id != 21 && value_id != 46 && value_id != 17 &&
-                value_id != 45 && value_id != 12 && value_id != 44){
-                    return false;
-                }
-            }else{
-                if(value_id != 0 && value_id != 56 && value_id != 55 &&
-                value_id != 31 && value_id != 54 && value_id != 24 &&
-                value_id != 53 && value_id != 23 && value_id != 27 &&
-                value_id != 22 && value_id != 20 && value_id != 18 &&
-                value_id != 15 && value_id != 13 && value_id != 10 &&
-                value_id != 8){
-                    return false;
-                }
-            }
-        }
-    }
-    else if(setting_id == PHOTO_LENS_ID){
-        int32_t value_id = PHOTO_LENS_VALUE[value_index];
-        if(value_id != 31 && value_id != 32) return false;
-    }
-    return true;
-}
